requirements.txt
fastapi==0.135.1
python-dotenv==1.0.1
uvicorn==0.35.0
